+++ service/controller/resource/tccpf/template/params_main_record_sets.go
@@ -1,8 +1,9 @@
 package template
 
 type ParamsMainRecordSets struct {
-	BaseDomain                 string
-	ClusterID                  string
-	GuestHostedZoneNameServers string
-	Route53Enabled             bool
+	BaseDomain                     string
+	ClusterID                      string
+	ControlPlanePublicHostedZoneID string
+	GuestHostedZoneNameServers     string
+	Route53Enabled                 bool
 }